投资顾问 获客信息 获客销售明细信息编辑删除代码提交

+++ src/main/java/com/baomidou/springwind/controller/AdvisorController.java
@@ -113,7 +113,7 @@ public class AdvisorController extends BaseController {
     @ResponseBody
     @Permission("9001")
     @RequestMapping("/delOne/{id}")
-    public String delUser(@PathVariable Long id) {
+    public String delUser(@PathVariable Integer id) {
         Boolean rlt = advisorService.deleteById(id);
         return rlt.toString();
     }
